+++ Server/src/main/java/com/bartlomiejskura/mymemories/controller/MemoryController.java
@@ -4,6 +4,8 @@ import com.bartlomiejskura.mymemories.exception.EntityNotFoundException;
 import com.bartlomiejskura.mymemories.model.Memory;
 import com.bartlomiejskura.mymemories.service.MemoryService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
@@ -18,12 +20,13 @@ public class MemoryController {
     private MemoryService memoryService;
 
     @GetMapping("/getAll")
-    public List<Memory> getAll(@RequestParam(value = "userId", required = false) Long userId, @RequestParam(value = "categoryId", required = false) Long categoryId){
-        if (userId != null) {
+    @PreAuthorize("#email.equals(authentication.name)")
+    public List<Memory> getAll(@RequestParam(value = "email", required = false) String email, @RequestParam(value = "categoryId", required = false) Long categoryId){
+        if (email != null) {
             if(categoryId != null){
-                return memoryService.getAllForUserAndCategory(userId, categoryId);
+                return memoryService.getAllForUserAndCategory(email, categoryId);
             }
-            return memoryService.getAllForUser(userId);
+            return memoryService.getAllForUser(email);
         }
         if(categoryId!=null){
             return memoryService.getAllForCategory(categoryId);
@@ -32,21 +35,24 @@ public class MemoryController {
     }
 
     @GetMapping("/getAllWithShared")
-    public List<Memory> getAllWithShared(@RequestParam(value = "userId") Long userId){
-        return Stream.concat(memoryService.getAllForUser(userId).stream(), memoryService.getAllSharedMemoriesForUser(userId).stream())
+    @PreAuthorize("#email.equals(authentication.name)")
+    public List<Memory> getAllWithShared(@RequestParam(value = "email") String email){
+        return Stream.concat(memoryService.getAllForUser(email).stream(), memoryService.getAllSharedMemoriesForUser(email).stream())
                 .collect(Collectors.toList());
     }
 
     @GetMapping("/getAllForDate")
-    public List<Memory> getAllForDateWithShared(@RequestParam(value = "userId") Long userId, @RequestParam(value = "time") String time){
+    @PreAuthorize("#email.equals(authentication.name)")
+    public List<Memory> getAllForDateWithShared(@RequestParam(value = "email") String email, @RequestParam(value = "time") String time){
         LocalDateTime date = LocalDateTime.parse(time);
 
-        return Stream.concat(memoryService.getAllForUserAndDate(userId, date).stream(),
-                memoryService.getAllSharedMemoriesForUserAndDate(userId, date).stream())
+        return Stream.concat(memoryService.getAllForUserAndDate(email, date).stream(),
+                memoryService.getAllSharedMemoriesForUserAndDate(email, date).stream())
                 .collect(Collectors.toList());
     }
 
     @PostMapping
+    @PreAuthorize("#memory.memoryOwner.email.equals(authentication.name)")
     public Memory addMemory(@RequestBody Memory memory){
         return memoryService.addMemory(memory);
     }
@@ -54,51 +60,59 @@ public class MemoryController {
     @GetMapping
     public Memory getMemory(@RequestParam(name="memoryId")Long memoryId){
         try{
-            return memoryService.getMemory(memoryId);
+            Memory memory = memoryService.getMemory(memoryId);
+            if(memory.getMemoryOwner().getEmail().equals(SecurityContextHolder.getContext().getAuthentication().getName())){
+                return memory;
+            }
+            return null;
         }catch (EntityNotFoundException e){
             return null;
         }
     }
 
     @PutMapping
+    @PreAuthorize("#memory.memoryOwner.email.equals(authentication.name)")
     public Memory editMemory(@RequestBody Memory memory){
         return memoryService.editMemory(memory);
     }
 
     @DeleteMapping
     public void deleteMemory(@RequestParam(name="memoryId")Long memoryId){
-        memoryService.deleteMemory(memoryId);
+        try{
+            if(memoryService.getMemory(memoryId).getMemoryOwner().getEmail().equals(SecurityContextHolder.getContext().getAuthentication().getName())){
+                memoryService.deleteMemory(memoryId);
+            }
+        }catch (EntityNotFoundException ignored){
+
+        }
     }
 
     @GetMapping("/getShared")
-    public List<Memory> getSharedMemories(@RequestParam(value = "userId", required = false) Long userId, @RequestParam(value = "categoryId", required = false) Long categoryId){
-        if (userId != null) {
-            return memoryService.getAllSharedMemoriesForUser(userId);
-        }
-        return null;
+    @PreAuthorize("#email.equals(authentication.name)")
+    public List<Memory> getSharedMemories(@RequestParam(value = "email") String email){
+        return memoryService.getAllSharedMemoriesForUser(email);
     }
 
     @GetMapping("/getFriendsPublicMemories")
-    public List<Memory> getFriendsPublicMemories(@RequestParam(value = "userId", required = false) Long userId){
-        if (userId != null) {
-            return memoryService.getFriendsPublicMemories(userId);
-        }
-        return null;
+    @PreAuthorize("#email.equals(authentication.name)")
+    public List<Memory> getFriendsPublicMemories(@RequestParam(value = "email") String email){
+        return memoryService.getFriendsPublicMemories(email);
     }
 
     @GetMapping("/search")
-    public List<Memory> search(@RequestParam(value = "userId") Long userId,
-                               @RequestParam(value = "keyword", required = false) String keyword,
-                               @RequestParam(value = "hasImage", required = false) Boolean hasImage,
-                               @RequestParam(value = "creationDateStart", required = false) String creationDateStart,
-                               @RequestParam(value = "creationDateEnd", required = false) String creationDateEnd,
-                               @RequestParam(value = "dateStart", required = false) String dateStart,
-                               @RequestParam(value = "dateEnd", required = false) String dateEnd,
-                               @RequestParam(value = "memoryPriorities", required = false) String memoryPriorities,
-                               @RequestParam(value = "publicToFriends", required = false) Boolean publicToFriends,
-                               @RequestParam(value = "isSharedMemory", required = false) Boolean isSharedMemory,
-                               @RequestParam(value = "categories", required = false) String categories) {
-        return memoryService.getMemories(userId, keyword, hasImage, creationDateStart==null?null:LocalDateTime.parse(creationDateStart),
+    @PreAuthorize("#email.equals(authentication.name)")
+    public List<Memory> search(@RequestParam(value = "email") String email,
+                                    @RequestParam(value = "keyword", required = false) String keyword,
+                                    @RequestParam(value = "hasImage", required = false) Boolean hasImage,
+                                    @RequestParam(value = "creationDateStart", required = false) String creationDateStart,
+                                    @RequestParam(value = "creationDateEnd", required = false) String creationDateEnd,
+                                    @RequestParam(value = "dateStart", required = false) String dateStart,
+                                    @RequestParam(value = "dateEnd", required = false) String dateEnd,
+                                    @RequestParam(value = "memoryPriorities", required = false) String memoryPriorities,
+                                    @RequestParam(value = "publicToFriends", required = false) Boolean publicToFriends,
+                                    @RequestParam(value = "isSharedMemory", required = false) Boolean isSharedMemory,
+                                    @RequestParam(value = "categories", required = false) String categories) {
+        return memoryService.getMemories(email, keyword, hasImage, creationDateStart==null?null:LocalDateTime.parse(creationDateStart),
                 creationDateEnd==null?null:LocalDateTime.parse(creationDateEnd),
                 dateStart==null?null:LocalDateTime.parse(dateStart), dateEnd==null?null:LocalDateTime.parse(dateEnd),
                 memoryPriorities, publicToFriends, isSharedMemory, categories);
